wip: update dir struct
