--- yas/src/info/const_info/mod.rs
@@ -1,8 +1,0 @@
-use super::*;
-
-mod genshin;
-mod starrail;
-
-type R = RectBound<WindowInfoType>;
-type P = Pos<WindowInfoType>;
-type S = Size<WindowInfoType>;
